feat: make highlight and category sub components, remove export file
