+++ src/pages/ui-components/Tag/Highlight/index.mdx
@@ -1,11 +1,11 @@
 ---
-name: TagHighlight
+name: Highlight
 ---
 
 import { Playground } from 'dokz';
 
-import Column from '../Grid/Column';
-import Row from '../Grid/Row';
+import Column from '../../Grid/Column';
+import Row from '../../Grid/Row';
 import TagHighlight from './';
 
 # TagHighlight
@@ -15,7 +15,7 @@ O componente TagHighlight é usado para destacar o status de um item.
 ### Importação
 
 ```js
-import TagHighlight from '@eduzz/houston-ui/TagHighlight';
+import TagHighlight from '@eduzz/houston-ui/Tag/Highlight';
 ```
 
 ### Exemplo
